chore: Update package dependencies
This commit updates the package.json file to include the latest versions of the "@neondatabase/serverless" and "@prisma/adapter-neon" dependencies. It also updates the "@prisma/client" dependency to version 5.18.0. Additionally, the devDependencies section includes the "@types/ws" dependency with version 8.5.12.

Note: This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes.